+++ src/Pim/Bundle/CatalogBundle/Doctrine/MongoDBODM/ProductRepository.php
@@ -758,4 +758,12 @@ class ProductRepository extends DocumentRepository implements
             );
         }
     }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getObjectManager()
+    {
+        return $this->getDocumentManager();
+    }
 }